Notification visualization and AJAX-updating
Added a notification button to the menu.
Added script to update notifications from "unread" to "read" using AJAX.
Conditions in the menu have been updated.
Added creation of some notifications (when the answer is updated and when the user does not have a CV if he tries to track the vacancy)